fix(arch): relocate switch-distinct-literals memo out of flow core.rs to satisfy size ratchet (#13598)
